Removed ActivitySourceAdapatee
